chore: rename to `dev`
